Use double calculations in planar path optimization

+++ source/MRMesh/MRPlanarPath.h
@@ -14,7 +14,7 @@ namespace MR
 /// \param outPath intermediate locations between s and e will be added here
 /// \param tmp elements will be temporary allocated here
 MRMESH_API bool reducePathViaVertex( const Mesh & mesh, const MeshTriPoint & start, VertId v, const MeshTriPoint & end, 
-    std::vector<MeshEdgePoint> & outPath, std::vector<Vector2f> & tmp );
+    std::vector<MeshEdgePoint> & outPath, std::vector<Vector2d> & tmp );
 
 /// reduces the length of given surface path: start-path-end
 MRMESH_API void reducePath( const Mesh & mesh, const MeshTriPoint & start, std::vector<MeshEdgePoint> & path, const MeshTriPoint & end, int maxIter = 5 );
